kube-apiserver/corerest: split apart generic code
